==> 00. data_read.ipynb <==
# Load Data

import pandas as pd
import numpy as np
from pandas import Series, DataFrame
import datetime as dt
from datetime import date
import pickle

#declare some variables
today = pd.Timestamp.today()
path_in = 'data/data_lokad'
path_out ='data/temp01'

## TODO 
# import reworks?

## Purchasing data

#load current file + format + add data source column
PURCHASING = pd.read_csv(path_in + '/LOKAD_PURCHASING.txt',delimiter='|',
                           dtype={
                               'PN':'str',
                               'PNM':'str',
                           })
PURCHASING['PURDATE'] = pd.to_datetime(PURCHASING['PURDATE'],format='%Y-%m-%d',errors='coerce')
PURCHASING['RECDATE'] = pd.to_datetime(PURCHASING['RECDATE'],format='%Y-%m-%d',errors='coerce')
PURCHASING['DATA_SOURCE'] = 'SQL'
#load Histo PO and format
HISTO_PO_ARO = pd.read_csv(path_in + '/HISTORICAL_PURCHASES_ARO.txt',delimiter='|',
                            dtype={'PNM':'str','item':'str'},
                            usecols=[1,2,3,4,5,6,7,8,9,10,11])
HISTO_PO_ARO = HISTO_PO_ARO.loc[HISTO_PO_ARO['QTYPENDING'] == 0] #delete suspicious open POs
HISTO_PO_ARO['DEPT_NAME'] = 'PURCHASING ARO'
HISTO_PO_ARO['DATA_SOURCE'] = 'HISTO ARO'
HISTO_PO_ARO['PURDATE'] = pd.to_datetime(HISTO_PO_ARO['PURDATE'], format='%m/%d/%Y %H:%M:%S')
HISTO_PO_ARO['RECDATE'] = pd.to_datetime(HISTO_PO_ARO['RECDATE'], format='%m/%d/%Y %H:%M:%S')

HISTO_PO_BIC = pd.read_csv(path_in + '/HISTORICAL_PURCHASES_BIC.txt',delimiter='|',
                          dtype={'PNM':'str'},
                          usecols=[1,2,3,4,5,6,7,8,9,10,11])
HISTO_PO_BIC = HISTO_PO_BIC.loc[HISTO_PO_BIC['QTYPENDING'] == 0] #delete suspicious open POs
HISTO_PO_BIC['DEPT_NAME'] = 'PURCHASING BIC'
HISTO_PO_BIC['DATA_SOURCE'] = 'HISTO BIC'
HISTO_PO_BIC['PURDATE'] = pd.to_datetime(HISTO_PO_BIC['PURDATE'], format='%m/%d/%Y %H:%M:%S')
HISTO_PO_BIC['RECDATE'] = pd.to_datetime(HISTO_PO_BIC['RECDATE'], format='%m/%d/%Y %H:%M:%S')

#append histo data
PURCHASING = PURCHASING.append(HISTO_PO_ARO, ignore_index = True, sort = False)
PURCHASING = PURCHASING.append(HISTO_PO_BIC, ignore_index = True, sort = False)

#cleanining
PURCHASING['QTYRECEIVED'] = pd.to_numeric(PURCHASING['QTYRECEIVED'], downcast='integer')
PURCHASING['QTYPENDING'] = pd.to_numeric(PURCHASING['QTYPENDING'], downcast='integer')
PURCHASING['EXT_COST'] = PURCHASING['COST'] * PURCHASING['QTYRECEIVED']
PURCHASING = PURCHASING[~PURCHASING['GEO_CODE'].isnull()] 
PURCHASING = PURCHASING[(PURCHASING['RECDATE'] <= today)|(PURCHASING['RECDATE']).isnull()] #remove POs received in the futur
PURCHASING['PN'] = PURCHASING['PN'].str.upper()
PURCHASING['PNM'] = PURCHASING['PNM'].str.replace('\.00','')


PURCHASING.to_pickle(path_out+'/PURCHASING.pkl')

#TODO: verif charts / numbers

## Stock Detail data

STOCK = pd.read_csv(path_in + '/LOKAD_STOCK_DETAIL.txt',delimiter='|',
                           dtype={
                               'PN':'str',
                               'PNM':'str'
                           })
STOCK['RECDATE'] = pd.to_datetime(STOCK['RECDATE'], format='%Y-%m-%d') #what to do the null ones??
#STOCK['UNIT_COST'] = STOCK['UNIT_COST'].str.replace('$','') # change the negative costs ((WTF))
#STOCK['UNIT_COST'] = STOCK['UNIT_COST'].str.replace('(','-')
#STOCK['UNIT_COST'] = STOCK['UNIT_COST'].str.replace(')','')
STOCK.drop(columns=['ID'],inplace = True)
STOCK['UNIT_COST'] = pd.to_numeric(STOCK['UNIT_COST'], downcast='float')
STOCK['PN'] = STOCK['PN'].str.upper()
STOCK['PNM'] = STOCK['PNM'].str.replace('\.00','')
STOCK.to_pickle(path_out + '/STOCK.pkl')

#TODO: verif charts / numbers

## Parts Master data

PARTS_MASTER = pd.read_csv(path_in + '/LOKAD_PARTS_MASTER.txt',delimiter='|',
                            dtype={
                                'PN':'str',
                                'PNM':'str',
                            })

PARTS_MASTER['LIST_PRICE'] = PARTS_MASTER['LIST_PRICE'].str.replace('$','') # change the negative costs ((WTF))
PARTS_MASTER['LIST_PRICE'] = PARTS_MASTER['LIST_PRICE'].str.replace('(','-')
PARTS_MASTER['LIST_PRICE'] = PARTS_MASTER['LIST_PRICE'].str.replace(')','')
PARTS_MASTER['PRICEDATE'] = pd.to_datetime(PARTS_MASTER['PRICEDATE'], format='%Y-%m-%d', errors='coerce')
PARTS_MASTER['LIST_PRICE'] = pd.to_numeric(PARTS_MASTER['LIST_PRICE'], downcast='float')
PARTS_MASTER['PN'] = PARTS_MASTER['PN'].str.upper()
PARTS_MASTER['STC'] = PARTS_MASTER['STC'].fillna('MISC')
PARTS_MASTER['CODE'] = PARTS_MASTER['CODE'].fillna('MISC')
PARTS_MASTER['CLASS'] = PARTS_MASTER['CLASS'].fillna('MISC')
PARTS_MASTER['PNM'] = PARTS_MASTER['PNM'].str.replace('\.00','')
PARTS_MASTER.to_pickle(path_out + '/PARTS_MASTER.pkl')
PARTS_MASTER.to_csv('PARTS_MAST.csv')
#TODO: verif charts / numbers

## Alternates data


ALTERNATES = pd.read_csv(path_in + '/LOKAD_ALTERNATES.txt',delimiter='|',
                            dtype={
                                'PNM':'str',
                                'ALT_PNM':'str',
                                'PN':'str',
                                'ALT_PN':'str'
                            })


ALTERNATES['PN'] = ALTERNATES['PN'].str.upper()
ALTERNATES['ATL_PN'] = ALTERNATES['ATL_PN'].str.upper()
ALTERNATES.rename(inplace = True, columns={'ATL_PN':'ALT_PN'})
ALTERNATES['PNM'] = ALTERNATES['PNM'].str.replace('\.00','')
ALTERNATES.to_pickle(path_out+'\ALTERNATES.pkl')


#TODO: verif charts / numbers

## Sales Orders data

#load current file
SALES_ORDERS = pd.read_csv(path_in + '/LOKAD_SALES_ORDERS.txt',delimiter='|',
                            dtype={
                                'PNM':'str',
                                'PN':'str'
                            })

SALES_ORDERS['DATA_SOURCE'] = 'SQL'
SALES_ORDERS['ENTRY_DATE'] = pd.to_datetime(SALES_ORDERS['ENTRY_DATE'], format='%Y-%m-%d', errors='coerce')
SALES_ORDERS['SHIP_DATE'] = pd.to_datetime(SALES_ORDERS['SHIP_DATE'], format='%Y-%m-%d', errors='coerce')


#load histo SO and format
HISTO_SO_BIC = pd.read_csv(path_in + '/HISTORICAL_SALES_BIC.txt',delimiter='|',
                          dtype={'PNM':'str'})
HISTO_SO_BIC = HISTO_SO_BIC.drop(columns='ID')
HISTO_SO_BIC['ENTRY_DATE'] = pd.to_datetime(HISTO_SO_BIC['ENTRY_DATE'], format = '%m/%d/%Y %H:%M:%S' )
HISTO_SO_BIC['SHIP_DATE'] = pd.to_datetime(HISTO_SO_BIC['ENTRY_DATE'], format = '%m/%d/%Y %H:%M:%S' )
HISTO_SO_BIC['DATA_SOURCE'] = 'HISTO BIC'

#parse histo data
SALES_ORDERS = SALES_ORDERS.append(HISTO_SO_BIC, ignore_index = True, sort = False)

#clean data
SALES_ORDERS['PN'] = SALES_ORDERS['PN'].str.upper()
SALES_ORDERS['PNM'] = SALES_ORDERS['PNM'].str.replace('\.00','')

# save as pickle
SALES_ORDERS.to_pickle(path_out + '/SALES_ORDERS.pkl')


#TODO: verif charts / numbers

## WO Bom data

#Load Current file
WO_BOM = pd.read_csv(path_in + '/LOKAD_WO_BOM.txt',delimiter='|',
                            dtype={
                                'PNM':'str',
                                'PN':'str'
                            })
WO_BOM['ADD_DATE'] = pd.to_datetime(WO_BOM['ADD_DATE'],format='%Y-%m-%d',errors='coerce')
WO_BOM['ISSUE_DATE'] = pd.to_datetime(WO_BOM['ISSUE_DATE'],format='%Y-%m-%d',errors='coerce')
WO_BOM['DATA_SOURCE'] = 'SQL'
WO_BOM['EXT_COST'] = WO_BOM['QTY_ISSUED'] * WO_BOM['COST']

#load histo WOB and format
HISTO_WOB_ARO = pd.read_csv(path_in + '/HISTORICAL_WO_BOM_ARO.txt',delimiter='|',
                           dtype={'PNM':'str',
                                  'WONO':'str'
                                  })
HISTO_WOB_ARO = HISTO_WOB_ARO.drop(columns=['ID'])

HISTO_WOB_ARO['ADD_DATE'] = pd.to_datetime(HISTO_WOB_ARO['ADD_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOB_ARO['ISSUE_DATE'] = pd.to_datetime(HISTO_WOB_ARO['ISSUE_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOB_ARO['DATA_SOURCE'] = 'HISTO ARO'
HISTO_WOB_ARO['GEO_CODE'] = 'DOR' #shouldn't been needed later

HISTO_WOB_BIC = pd.read_csv(path_in + '/HISTORICAL_WO_BOM_BIC.txt',delimiter='|',
                           dtype={'PNM':'str','WONO':'str'})
HISTO_WOB_BIC['ADD_DATE'] = pd.to_datetime(HISTO_WOB_BIC['ADD_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOB_BIC['ISSUE_DATE'] = pd.to_datetime(HISTO_WOB_BIC['ISSUE_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOB_BIC = HISTO_WOB_BIC.drop(columns=['ID'])
HISTO_WOB_BIC['DATA_SOURCE'] = 'HISTO BIC'
HISTO_WOB_BIC['GEO_CODE'] = 'MIA MRO' #shouldn't been needed later

#Append histo to current
WO_BOM= WO_BOM.append(HISTO_WOB_ARO, ignore_index = True, sort = False)
WO_BOM= WO_BOM.append(HISTO_WOB_BIC, ignore_index = True, sort = False)

#Cleaning
WO_BOM = WO_BOM[~WO_BOM['ACTIVITY'].str.contains('2014')] #bad conversions from histo data
WO_BOM = WO_BOM[~WO_BOM['ACTIVITY'].str.contains('2015')]
WO_BOM = WO_BOM[~WO_BOM['ACTIVITY'].str.contains('2016')]
WO_BOM = WO_BOM.loc[WO_BOM['ACTIVITY'] != 'CLOSED']  #bad conversions from histo data
WO_BOM['ITEM'] = WO_BOM['ITEM'].str.upper()
WO_BOM.rename(inplace = True, columns={'ITEM':'PN'})
WO_BOM['PNM'] = WO_BOM['PNM'].str.replace('\.00','')


#save as pickly
WO_BOM.to_pickle(path_out + '/WO_BOM.pkl')

#TODO: verif charts / numbers

## WO Header data

#Load current file
WO_HEADER = pd.read_csv(path_in + '/LOKAD_WO_HEADER.txt',delimiter='|',
                            dtype={
                                'PNM':'str',
                                'PN':'str'
                            })
WO_HEADER['DATA_SOURCE'] = 'SQL'
WO_HEADER['ENTRY_DATE'] = pd.to_datetime(WO_HEADER['ENTRY_DATE'],format='%Y-%m-%d',errors='coerce')
WO_HEADER['CLOSE_DATE'] = pd.to_datetime(WO_HEADER['CLOSE_DATE'],format='%Y-%m-%d',errors='coerce')

#load histo WOH and format
HISTO_WOH_ARO = pd.read_csv(path_in + '/HISTORICAL_WO_HEADER_ARO.txt',delimiter='|',
                           dtype={'PNM':'str','WONO':'str'},
                           usecols=[1,2,3,4,5,6,7,8,9,10,11,12])
HISTO_WOH_ARO['ENTRY_DATE'] = pd.to_datetime(HISTO_WOH_ARO['ENTRY_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOH_ARO['CLOSE_DATE'] = pd.to_datetime(HISTO_WOH_ARO['CLOSE_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOH_ARO['DATA_SOURCE'] = 'HISTO ARO'
HISTO_WOH_ARO['GEO_CODE'] = np.where(HISTO_WOH_ARO['GEO_CODE'] == 'MED','MED MRO',HISTO_WOH_ARO['GEO_CODE'])

HISTO_WOH_BIC = pd.read_csv(path_in + '/HISTORICAL_WO_HEADER_BIC.txt',delimiter='|',
                           dtype={'PNM':'str','WONO':'str'})
HISTO_WOH_BIC = HISTO_WOH_BIC.drop(columns=['ID'])
HISTO_WOH_BIC['ENTRY_DATE'] = pd.to_datetime(HISTO_WOH_BIC['ENTRY_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOH_BIC['CLOSE_DATE'] = pd.to_datetime(HISTO_WOH_BIC['CLOSE_DATE'],  format='%m/%d/%Y %H:%M:%S')
HISTO_WOH_BIC['DATA_SOURCE'] = 'HISTO BIC'


#append the 2 histo files together
HISTO_WOH = HISTO_WOH_BIC.append(HISTO_WOH_ARO, ignore_index = True, sort = False)

#drop all records from histo if WO also exists in current WO_HEADER
HISTO_WOH = HISTO_WOH[~HISTO_WOH['WONO'].isin(WO_HEADER['WONO'].unique())]

#append histoical data to current data
WO_HEADER= WO_HEADER.append(HISTO_WOH, ignore_index = True, sort = False)

#clean data
WO_HEADER = WO_HEADER.loc[WO_HEADER['GEO_CODE'] != 'TEST']
WO_HEADER = WO_HEADER.loc[~WO_HEADER['GEO_CODE'].isnull()]
WO_HEADER.drop_duplicates(subset=['WONO','SERIAL_NUMBER'],keep='last',inplace = True)
WO_HEADER['PN'] = WO_HEADER['PN'].str.upper()
WO_HEADER['PNM'] = WO_HEADER['PNM'].str.replace('\.00','')

#save as pickle
WO_HEADER.to_pickle(path_out + '/WO_HEADER.pkl')

#TODO: verif charts / numbers

## Capabilities data

CAPABILITIES = pd.read_csv(path_in + '/BA_VIEW_INTERNAL_CAPABILTIES.txt',delimiter='|')
CAPABILITIES['IS_ACTIVE'] = np.where(CAPABILITIES['CAPABILITY_CODE'].str.contains('INACTIVE'),
                                     '0', '1') #adds colunm to show active caps

#CAPABILITIES
CAPABILITIES = CAPABILITIES[CAPABILITIES['CAPABILITY_CODE'] != 'Barfield, Inc.']
CAPABILITIES = CAPABILITIES[~CAPABILITIES['CAPABILITY_CODE'].isnull()]
CAPABILITIES = CAPABILITIES.drop(columns = ['APPLICATION_CODE','DESCRIPTION',
                                            'MFG_CODE','MFG_NAME','STOCK_CATEGORY_CODE',
                                            'CAPABILITY_CODE','EST_COST','ACTIVE_MANUALS',
                                            'SYSCM_AUTO_KEY','SYCPC_AUTO_KEY','INACTIVE_MANUALS'
                                           ])
CAPABILITIES.rename(columns = {'PNM_AUTO_KEY':'PNM'}, inplace = True)
CAPABILITIES['IS_ACTIVE'] = pd.to_numeric(CAPABILITIES['IS_ACTIVE'], downcast='integer')


#sum "active" field in case they exist in both condition - by default show them active
CAPABILITIES = pd.pivot_table(CAPABILITIES, values ='IS_ACTIVE', aggfunc=np.sum,index=['PN','GEO_CODE']).reset_index()
CAPABILITIES['IS_ACTIVE'] = np.where(CAPABILITIES['IS_ACTIVE']>=1, 1, 0)

#clean
CAPABILITIES['PN'] = CAPABILITIES['PN'].str.upper()

#save as pickle
CAPABILITIES.to_pickle(path_out + '/CAPABILITIES.pkl')




#TODO: verif charts / numbers

## All other Static files (not sure that this is useful here yet)

# --- Distribution parts
DISTRIB_PARTS = pd.read_csv(path_in + '/BICdistributedparts.csv',delimiter=';',
                                dtype={
                                    'PNM':'str',
                                    'PN':'str'
                                })
DISTRIB_PARTS['PN'] = DISTRIB_PARTS['PN'].str.upper()
DISTRIB_PARTS.drop(columns = ['DESCRIPTION','OEM','PNM'])
DISTRIB_PARTS.to_pickle(path_out + '/DISTRIB_PARTS.pkl')


# --- Excluded PN from Sales Orders
EXCLUD_PARTS_SO = pd.read_csv(path_in + '/ExcludedPartsForSalesOrders.csv',
                              delimiter=',',
                              dtype={
                                     'PN':'str' 
                              })
EXCLUD_PARTS_SO['PN'] = EXCLUD_PARTS_SO['PN'].str.upper()
EXCLUD_PARTS_SO['DateAdded'] = pd.to_datetime(EXCLUD_PARTS_SO['DateAdded'])
EXCLUD_PARTS_SO.to_pickle(path_out + '/EXCLUD_PARTS_SO.pkl')

# --- Excluded Customer from Sales Orders
EXCLUD_CLIENT_SO = pd.read_csv(path_in + '/ExcludedClientsForSalesOrders.csv',
                               delimiter=',')
EXCLUD_CLIENT_SO['DateAdded'] = pd.to_datetime(EXCLUD_CLIENT_SO['DateAdded'])
EXCLUD_CLIENT_SO.to_pickle(path_out + '/EXCLUD_CLIENT_SO.pkl')

# --- Excluded SO
EXCLUD_SO = pd.read_csv(path_in + '/ExcludedSalesOrders.csv',delimiter=',')
EXCLUD_SO['DateAdded'] = pd.to_datetime(EXCLUD_SO['DateAdded'])
EXCLUD_SO.to_pickle(path_out + '/EXCLUD_SO.pkl')

# --- Excluded parts from All Consumption
EXCLUDED_PARTS = pd.read_csv(path_in + '/PartsToBeExcluded.csv',delimiter=',')
EXCLUDED_PARTS.drop(columns = 'Description', inplace = True)
EXCLUDED_PARTS['date added'] = pd.to_datetime(EXCLUDED_PARTS['date added'])
EXCLUDED_PARTS.rename(columns={'item':'PN','GEO_CODE_UPPER':'UPPER_GEO_CODE'},inplace=True)
EXCLUDED_PARTS['PN'] = EXCLUDED_PARTS['PN'].str.upper()
EXCLUDED_PARTS.to_pickle(path_out + '/EXCLUDED_PARTS.pkl')

# --- PMAS ACCEPTANCE
PMA_ACCEPTANCE = pd.read_csv(path_in + '/pma_acceptance.csv',delimiter=',',
                            dtype={'PN':'str'})
PMA_ACCEPTANCE.rename(columns={'item':'LRU_PN','item':'PN'},inplace=True)
PMA_ACCEPTANCE['PN'] = PMA_ACCEPTANCE['PN'].str.upper()
PMA_ACCEPTANCE.to_pickle(path_out + '/PMA_ACCEPTANCE.pkl')

    
# --- VIPS
VIP_LRU = pd.read_csv(path_in + '/VIP_LRUs.csv',delimiter=',')
VIP_LRU.rename(columns = {'LRU':'PN'}, inplace= True)
VIP_LRU['PN'] = VIP_LRU['PN'].str.upper()
VIP_LRU['Date_entered'] = pd.to_datetime(VIP_LRU['Date_entered'])
VIP_LRU.to_pickle(path_out + '/VIP_LRU.pkl')


# --- WON BIZ
WON_BIZ = pd.read_csv(path_in + '/WON_BUSINESS.csv',delimiter=',')
WON_BIZ.to_pickle(path_out + '/WON_BIZ.pkl')


# --- LOST BIZ
LOST_BIZ = pd.read_csv(path_in + '/LOST_BUSINESS.csv',delimiter=',')
LOST_BIZ.rename(columns = {'part number':'PN','Customer':'COMPANY_CODE','GEO_CODE_UPPER':'UPPER_GEO_CODE'}, inplace= True)
LOST_BIZ['PN'] = LOST_BIZ['PN'].str.upper()
LOST_BIZ['Date entered'] = pd.to_datetime(LOST_BIZ['Date entered'])
LOST_BIZ['Termination date'] = pd.to_datetime(LOST_BIZ['Termination date'])
LOST_BIZ.to_pickle(path_out + '/LOST_BIZ.pkl')



MANUAL_MIN_MAX = pd.read_csv(path_in + '/min_max_static.csv',delimiter=',',
                            dtype={
                                'PN':'str'
                            })

WHS = pd.read_csv(path_in + '/WHS.csv',delimiter=',') # this one shouldn't be static -
                                                      #need to SQL it as well as the DEPT one
